+++ lambda-runtime/src/test/scala/feral/lambda/runtime/LambdaRuntimeSuite.scala
@@ -22,75 +22,15 @@ import org.http4s.circe.jsonEncoderWithPrinter
 
 import scala.concurrent.duration.DurationInt
 
-class LambdaRuntimeSuite extends CatsEffectSuite {
-
-  implicit val jsonEncoder: EntityEncoder[IO, Json] = jsonEncoderWithPrinter[IO](Printer.noSpaces.copy(dropNullValues = true))
-
-  def createTestEnv(funcName: IO[String] = IO("test"),
-                    memorySize: IO[Int] = IO(144),
-                    funcVersion: IO[String] = IO("1.0"),
-                    logGroupName: IO[String] = IO("test"),
-                    logStreamName: IO[String] = IO("test"),
-                    runtimeApi: IO[String] = IO("testApi")
-                   ): LambdaRuntimeEnv[IO] = new LambdaRuntimeEnv[IO] {
-    override def lambdaFunctionName: IO[String] = funcName
-
-    override def lambdaFunctionMemorySize: IO[Int] = memorySize
-
-    override def lambdaFunctionVersion: IO[String] = funcVersion
-
-    override def lambdaLogGroupName: IO[String] = logGroupName
-
-    override def lambdaLogStreamName: IO[String] = logStreamName
-
-    override def lambdaRuntimeApi: IO[String] = runtimeApi
-  }
-
-  def defaultRoutes(invocationQuota: Ref[IO, Int]): HttpRoutes[IO] = HttpRoutes.of[IO] {
-    case GET -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / "next" =>
-      for {
-        currentInvocations <- invocationQuota.modify(cur => (cur-1, cur))
-        _ <- if (currentInvocations > 0) IO.unit else IO.never // infinite loop if not first request, assumed behaviour of actual next invocation endpoint?
-        headers = Headers(
-          `Lambda-Runtime-Aws-Request-Id`("testId"),
-          `Lambda-Runtime-Deadline-Ms`(20),
-          `Lambda-Runtime-Invoked-Function-Arn`("test"),
-          `Lambda-Runtime-Client-Context`(new ClientContext(
-            new ClientContextClient(
-              "test",
-              "test",
-              "test",
-              "test",
-              "test"
-            ),
-            new ClientContextEnv(
-              "test",
-              "test",
-              "test",
-              "test",
-              "test"
-            ),
-            JsonObject.empty
-          )),
-          `Lambda-Runtime-Client-Identity`(new CognitoIdentity("test", "test"))
-        )
-        body = Json.obj("eventField" -> "test".asJson)
-        resp <- Ok(body, headers)
-      } yield resp
-    case POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "response" => Ok()
-    case POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "error" => Ok()
-  }
+class LambdaRuntimeSuite extends BaseRuntimeSuite {
 
   test("The runtime can process an event and pass the result to the invocation response url") {
     implicit val env: LambdaRuntimeEnv[IO] = createTestEnv()
     for {
       invocationQuota <- Ref[IO].of(1)
       eventualInvocationId <- Deferred[IO, String]
-      client = Client.fromHttpApp[IO]((HttpRoutes.of[IO] {
-        case POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "response" =>
-          eventualInvocationId.complete(id) >> Ok()
-      } <+> defaultRoutes(invocationQuota)).orNotFound)
-      handler = (_: Json, _: Context[IO]) => JsonObject.empty.asJson.pure[IO]
+      client = Client.fromHttpApp[IO]((testNextInvocationRoute(eventualInvocationId) <+> defaultRoutes(invocationQuota)).orNotFound)
+      handler = (_: Json, _: Context[IO]) => Json.obj().pure[IO]
       runtimeFiber <- FeralLambdaRuntime(client)(Resource.eval(handler.pure[IO])).start
       invocationId <- eventualInvocationId.get.timeout(2.seconds)
       _ <- runtimeFiber.cancel
@@ -103,15 +43,15 @@ class LambdaRuntimeSuite extends CatsEffectSuite {
       invocationQuota <- Ref[IO].of(1)
       eventualInvocation <- Deferred[IO, (Json, Context[IO])]
       client = Client.fromHttpApp[IO](defaultRoutes(invocationQuota).orNotFound)
-      handler = (json: Json, context: Context[IO]) => eventualInvocation.complete((json, context)) >> JsonObject.empty.asJson.pure[IO]
+      handler = (json: Json, context: Context[IO]) => eventualInvocation.complete((json, context)) >> Json.obj().pure[IO]
       runtimeFiber <- FeralLambdaRuntime(client)(Resource.eval(handler.pure[IO])).start
       (jsonEvent, context) <- eventualInvocation.get.timeout(2.seconds)
       expectedJson = Json.obj("eventField" -> "test".asJson)
       _ <- runtimeFiber.cancel
     } yield {
       assert(jsonEvent eqv expectedJson)
-      assert(context.clientContext.get.client.appTitle == "test")
-      assert(context.identity.get.identityId == "test")
+      assert(context.clientContext.exists(_.client.appTitle == "test"))
+      assert(context.identity.exists(_.identityId == "test"))
       assert(context.functionName == "test")
       assert(context.memoryLimitInMB == 144)
     }
@@ -126,24 +66,12 @@ class LambdaRuntimeSuite extends CatsEffectSuite {
     for {
       invocationQuota <- Ref[IO].of(1)
       eventualInvocationError <- Deferred[IO, Json]
-      client = Client.fromHttpApp((HttpRoutes.of[IO] {
-        case req@POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "error" =>
-          for {
-            body <- req.body.compile.to(Array).flatMap(Parser.parseFromByteArray(_).liftTo[IO])
-            _ <- eventualInvocationError.complete(body)
-            resp <- Ok()
-          } yield resp
-      } <+> defaultRoutes(invocationQuota)).orNotFound)
+      client = Client.fromHttpApp((testInvocationErrorRoute(eventualInvocationError) <+> defaultRoutes(invocationQuota)).orNotFound)
       handler = (_: Json, _: Context[IO]) => IO.raiseError[Json](new Exception("oops"))
       runtimeFiber <- FeralLambdaRuntime(client)(Resource.eval(handler.pure[IO])).start
-      invocationError <- eventualInvocationError.get.timeout(2.seconds)
-      expectedBody = Json.obj(
-        "errorMessage" -> "oops".asJson,
-        "errorType" -> "exception".asJson,
-        "stackTrace" -> List[String]().asJson
-      )
+      errorRequest <- eventualInvocationError.get.timeout(2.seconds)
       _ <- runtimeFiber.cancel
-    } yield assert(invocationError eqv expectedBody)
+    } yield assert(errorRequest eqv invocationErrorRequest())
   }
 
   test("The runtime will call the invocation error url when a needed environment variable is not available") {
@@ -151,59 +79,35 @@ class LambdaRuntimeSuite extends CatsEffectSuite {
     for {
       invocationQuota <- Ref[IO].of(1)
       eventualInvocationError <- Deferred[IO, Json]
-      client = Client.fromHttpApp((HttpRoutes.of[IO] {
-        case req@POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "error" =>
-          for {
-            body <- req.body.compile.to(Array).flatMap(Parser.parseFromByteArray(_).liftTo[IO])
-            _ <- eventualInvocationError.complete(body)
-            resp <- Ok()
-          } yield resp
-      } <+> defaultRoutes(invocationQuota)).orNotFound)
-      handler = (_: Json, _: Context[IO]) => JsonObject.empty.asJson.pure[IO]
+      client = Client.fromHttpApp((testInvocationErrorRoute(eventualInvocationError) <+> defaultRoutes(invocationQuota)).orNotFound)
+      handler = (_: Json, _: Context[IO]) => Json.obj().pure[IO]
       runtimeFiber <- FeralLambdaRuntime(client)(Resource.eval(handler.pure[IO])).start
-      invocationError <- eventualInvocationError.get.timeout(2.seconds)
-      expectedBody = Json.obj(
-        "errorMessage" -> "oops".asJson,
-        "errorType" -> "exception".asJson,
-        "stackTrace" -> List[String]().asJson
-      )
+      errorRequest <- eventualInvocationError.get.timeout(2.seconds)
       _ <- runtimeFiber.cancel
-    } yield assert(invocationError eqv expectedBody)
-
+    } yield assert(errorRequest eqv invocationErrorRequest())
   }
 
-  // TODO needs refactoring
   test("The runtime will continue to process events even after encountering an invocation error") {
     implicit val env: LambdaRuntimeEnv[IO] = createTestEnv()
     for {
       invocationQuota <- Ref[IO].of(2)
       eventualInvocationError <- Deferred[IO, Json]
       eventualResponse <- Deferred[IO, String]
-      client = Client.fromHttpApp((HttpRoutes.of[IO] {
-        case req@POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "error" =>
-          for {
-            body <- req.body.compile.to(Array).flatMap(Parser.parseFromByteArray(_).liftTo[IO])
-            _ <- eventualInvocationError.complete(body)
-            resp <- Ok()
-          } yield resp
-        case POST -> Root / "testApi" / FeralLambdaRuntime.ApiVersion / "runtime" / "invocation" / id / "response" =>
-          eventualResponse.complete(id) >> Ok()
-      } <+> defaultRoutes(invocationQuota)).orNotFound)
+      client = Client.fromHttpApp(
+        (testInvocationErrorRoute(eventualInvocationError)
+        <+> testNextInvocationRoute(eventualResponse)
+        <+> defaultRoutes(invocationQuota)).orNotFound
+      )
       handler = (_: Json, _: Context[IO]) => for {
         currentInvocations <- invocationQuota.get
-        resp <- if (currentInvocations < 1) JsonObject.empty.asJson.pure[IO] else IO.raiseError(new Exception("oops first invocation"))
+        resp <- if (currentInvocations < 1) Json.obj().pure[IO] else IO.raiseError(new Exception("oops first invocation"))
       } yield resp
       runtimeFiber <- FeralLambdaRuntime(client)(Resource.eval(handler.pure[IO])).start
       invocationError <- eventualInvocationError.get.timeout(2.seconds)
-      expectedErrorBody = Json.obj(
-        "errorMessage" -> "oops first invocation".asJson,
-        "errorType" -> "exception".asJson,
-        "stackTrace" -> List[String]().asJson
-      )
       secondInvocationResponse <- eventualResponse.get.timeout(2.seconds)
       _ <- runtimeFiber.cancel
     } yield {
-      assert(invocationError eqv expectedErrorBody)
+      assert(invocationError eqv invocationErrorRequest("oops first invocation"))
       assert(secondInvocationResponse eqv "testId")
     }
   }
